UPSTREAM: soc/intel/skylake: Fix broken memory info HOB scanning
It looks like this code was written with completely different semantics
in mind. Controllers, channels and DIMMs are all presented in their phy-
sical order (i.e. gaps are not closed). So we have to look at the whole
structure and not only the first n respective entries.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ib848908050be6d534eb7314842cd3d14a120dbcb
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 9dc62ea133d826cf88ccd29e579ea976257d5487
Original-Change-Id: I8a9039f73f1befdd09c1fc8e17cd3f6e08e0cd47
Original-Signed-off-by: Nico Huber <nico.huber@secunet.com>
Original-Reviewed-on: https://review.coreboot.org/20650
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/579914
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>