UPSTREAM: util/abuild: Set exit status on failure
Currently, the what-jenkins-does target doesn't stop if abuild fails.
This change gives a command line parameter to abuild with tells it to
exit with a non-zero exit status if any of the builds failed.

This will be particularly useful for concourse which doesn't parse the
junit.xml files, and only knows whether a build passed or failed by the
exit status.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ia70de90721866a3e4883504a34bafe6e877d6c3c
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: ba973bd2de8bf91cc83431333519124bf0f1fd72
Original-Change-Id: Ic2be5deaedacd3f55db8e0b14a2b7b39cc44664e
Original-Signed-off-by: Martin Roth <martinroth@google.com>
Original-Reviewed-on: https://review.coreboot.org/20656
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Chris Ching <chingcodes@google.com>
Reviewed-on: https://chromium-review.googlesource.com/595282
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>