UPSTREAM: northbridge/intel/i440bx: Merge RAM init routines
There are 4 routines used in RAM init that most if not all
i440bx mainboards call in the same order. Implements a single
RAM init routine for them to allow for future consolidation.
Boards to be changed to use this one routine in a future change.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ieaf850e2c6808f11e684bcc4fb018853d3c11747
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 078e324080901f509493a08c7bf6d28e96119e3f
Original-Change-Id: Ib553b07b117de12b7982586bce0f9355f55013a0
Original-Signed-off-by: Keith Hui <buurin@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/20676
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Martin Roth <martinroth@google.com>
Reviewed-on: https://chromium-review.googlesource.com/582853
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>