UPSTREAM: libpayload: video: Add support for font scaling with a factor
This introduces support for font scaling with a factor provided via
Kconfig. In practice, the font itself is not scaled at any point in
memory and only the logic to determine whether a pixel should be filled
or not is changed.

Thus, it should not significantly impact either the access time or
memory use.

BUG=none
BRANCH=none
TEST=none

Change-Id: I9d9e8495990d777a1fa6b3b615c91d2479b4b946
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: d85e485c5892f2d96e8c5d10828c13af154a5481
Original-Change-Id: Idff210617c9ec08c6034aef107cfdb34c7cdf029
Original-Signed-off-by: Paul Kocialkowski <contact@paulk.fr>
Original-Reviewed-on: https://review.coreboot.org/20709
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/608463
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>