UPSTREAM: copy & update test routines from what-jenkins-does
Currently the only testing we had was 'what-jenkins-does' and
'make lint'. While the lint testing is suitable for developers,
the 'what-jenkins-does' target really isn't, as it was designed
specifically for testing on jenkins.

This adds the infrastructure for basic tests that are more suitable
for the developer. Extended tests and improvements will follow.

Add the coreboot-builds directories to .gitignore.

TODO:
- Save/restore .config
- Update test-abuild to use existing COREBOOT_BUILD_DIR variable

BUG=none
BRANCH=none
TEST=none

Change-Id: Ib8116dedb9eb534ff808e0603be301c8d60875ec
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 6c9c196f1cdf5e1a94c4ae312318dc7934508023
Original-Change-Id: I19e1256d79531112ff84e47a307f55791533806f
Original-Signed-off-by: Martin Roth <martinroth@google.com>
Original-Reviewed-on: https://review.coreboot.org/20874
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Chris Ching <chingcodes@google.com>
Reviewed-on: https://chromium-review.googlesource.com/612358
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>