UPSTREAM: AGESA f15 vendorcode: Remove AM3r2 refcode
We never had a board in the tree that implements this.

If you are interested in implementing such board, note
that also f12 and f14 had copies of the same refcode.
As part of the sourcetree cleanup it was not studied
which was the most up-to-date one for AM3r2.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ia4952bf2831d38ba453b2b0f7ca4800bb98b4aee
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 529ef49590664478c2d4e7118271070a33620b7f
Original-Change-Id: Ic7dd065c0df08c22af6f3a2dcfc7ff47d6283a46
Original-Signed-off-by: Kysti Mlkki <kyosti.malkki@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/21255
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-by: Patrick Georgi <pgeorgi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/642795
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>