UPSTREAM: mainboard/lippert: Refactor SEMA watchdog message
It's too critical to ignore when sending the message on
SMBus fails, so allow for a fair amount of retries.
Failure here causes watchdog to do hard reset later.

Move it out of mainboard.c as we need to call this
early in romstage while we are debugging.

BUG=none
BRANCH=none
TEST=none

Change-Id: I80d99d30082f444ce5f4ef3073978e83d0a084f8
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 081b66951f04bb67ad2d04750212a45fc0c24e90
Original-Change-Id: I1006b079269d6dd44de630db7a5694124af2f974
Original-Signed-off-by: Kysti Mlkki <kyosti.malkki@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/21316
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/657127
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>