UPSTREAM: board_status: Tell user where to find output when results are uploaded
If results are uploaded the temporary directory in which they are
stored gets deleted, yet we currently point to the deleted directory in
the output.

This patch fixes it so that we point to the actual location in the local
repository where uploaded results are found.

BUG=none
BRANCH=none
TEST=none

Change-Id: I620a9ee3750d7f0dd6d500978b09430d182fb270
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 588a72210d5e7cbefc46eef4b3b1d1a48bfeff8e
Original-Change-Id: I1f42c3296ec1d19fcfa4911307e07e67de289895
Original-Signed-off-by: David Hendricks <david.hendricks@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/21415
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Martin Roth <martinroth@google.com>
Reviewed-on: https://chromium-review.googlesource.com/675107
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>