UPSTREAM: mb/intel/dg43gt: Fix smbus IRQ
This board uses the reset defaults for DxxIP and DxxIR.
The datasheet "Intel  I/O Controller Hub 10 (ICH10) Family"
mistakenly says in the D31IP register that all function have INTB as
default. This is however not true as documented in the reset default
value.

This fixes the DSDT such that the SMBus device gets a route for the
INT C interrupt it uses.

BUG=none
BRANCH=none
TEST=none

Change-Id: If11c3e9f62f81679bc72e42fd62e04cd31af996b
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 92ce1fb45eda81d8667db63ec8eb649ed87c743c
Original-Change-Id: I3dd1308fb7acec86b90ecd9d2079cf9a58702c40
Original-Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-on: https://review.coreboot.org/21442
Original-Reviewed-by: Kysti Mlkki <kyosti.malkki@gmail.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/666765
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>