UPSTREAM: sb/intel/bd82x6x: Add awareness of ME's Alt Disable Mode
me_cleaner now allows setting a bit in the PCH straps - AltMeDisable
tells the ME to stop execution after BUP - disabling the 30 minute
watchdog - but also "breaking" the ME. The ME reports opmode = 2.

This means the ME will not respond when we wait for an acknowledgement
about the DRAM being ready. The current code waits 5 seconds for a
response, that in this case, never comes.

If the ME is reporting opmode 2, don't delay or wait for a response
from the ME.

Tested on my X220, this patch fixed the five seconds before the payload
executed. Verified using the timestamp patch.

BUG=none
BRANCH=none
TEST=none

Change-Id: Id82737c0bb37decbf945be27b6f7a3c71c1ecba0
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: d7e0cb93ae8ead42aa20dd4dcafa137fc28653bc
Original-Change-Id: Ifdda6b2dbb8ae3a650be6d5df6c60475a3fa74aa
Original-Signed-off-by: Nathaniel Roach <nroach44@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/21466
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
Reviewed-on: https://chromium-review.googlesource.com/671345
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>