UPSTREAM: board_status: Make board_status more friendly for local usage
board_status.sh was originally written for use cases where the DUT
is remote, i.e. accessed via serial port or SSH. This lead to some
issues when attempting to run the script on the DUT itself.

This patch attempts to handle the local use case more gracefully.
sudo is used when running the cbmem command, and the '-c' option
can be used to set cbmem path in case it's not in the default path
used by sudo.

BUG=none
BRANCH=none
TEST=none

Change-Id: I1396e8ee4e38f6b1923deeb8610c3a90e4a25a2c
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: bb90fb55d0bd09ad2865ed791641f1fcf778f0eb
Original-Change-Id: I62957678ccae65fc46fd6ddf5ae92983d36cffad
Original-Signed-off-by: David Hendricks <david.hendricks@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/21566
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Martin Roth <martinroth@google.com>
Reviewed-on: https://chromium-review.googlesource.com/924237
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>