UPSTREAM: soc/amd/common: Add included directory
If the symbol SOC_AMD_COMMON is selected, include the soc/amd/common
directory.  Until now this has been working due to the directory being
included as part of AGESA_INC in vendorcode.  That one is still
necessary in order to build the AGESA code so it is left in place
for now.

BUG=none
BRANCH=none
TEST=none

Change-Id: I684e668a3dfae0eb2a0113461e2927bc0fcb941e
Signed-off-by: Martin Roth <martinroth@google.com>
Original-Commit-Id: b639bef83644a13ae8864095c8635c722bd37754
Original-Change-Id: Ia8191897d2030c475c9268ae86faaf01952c6ace
Original-Signed-off-by: Marshall Dawson <marshalldawson3rd@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/21738
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/699373
Commit-Ready: Martin Roth <martinroth@chromium.org>
Tested-by: Martin Roth <martinroth@chromium.org>
Reviewed-by: Martin Roth <martinroth@chromium.org>