UPSTREAM: google/kahlee: Add defines in OemCustomize.c
Add a #define for MB_DIMM_SLOTS and verify it doesn't exceed the max
supported for the device.  AGESA's DRAM procedures follow the BKDG and
may vary depending on the number of slots on the motherboard.  DIMM
numbering and ordering is also affected by this value.

Replace hardcoded integers with defined values for DIMM slots and
number of channels.

BUG=none
BRANCH=none
TEST=none

Change-Id: I25206f8b1dd51643c6a791ab7510c5fbaaef8018
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 3f6c400310a930d226e85e77f29c51f03578f0f1
Original-Change-Id: I4f7336da80b4e3d7f351502a63de0652e9ff5395
Original-Signed-off-by: Marshall Dawson <marshalldawson3rd@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/21853
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/765337
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>