UPSTREAM: amd/gardenia: Add defines in OemCustomize.c
Add a #define for MB_DIMM_SLOTS and verify it doesn't exceed the max
supported for the device.  AGESA's DRAM procedures follow the BKDG and
may vary depending on the number of slots on the motherboard.  DIMM
numbering and ordering is also affected by this value.

Replace hardcoded integers with defined values for DIMM slots and
number of channels.

BUG=none
BRANCH=none
TEST=none

Change-Id: I0a639f3c022d27cc9ece8ad0e90d2265cc5ee3c4
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 209ea1459bd36191d7012d458f3f84e99a3d8ed2
Original-Change-Id: I81aa0165660e7627f1d977ac40479700cff8b80b
Original-Signed-off-by: Marshall Dawson <marshalldawson3rd@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/21854
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/778970
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>